refactor:  move enbironments
