+++ libs/root/environments/src/lib/energiDataServiceEndpoint.env.ts
