--- libs/root/environments/src/lib/root-environments.module.ts
@@ -1,7 +1,0 @@
-import {NgModule} from '@angular/core'
-import {CommonModule} from '@angular/common'
-
-@NgModule({
-  imports: [CommonModule],
-})
-export class RootEnvironmentsModule {}
